% Auto-generated: do not edit by hand
\name{checkTreePicker}

\alias{checkTreePicker}

\title{CheckTreePicker component}

\description{

}

\usage{
checkTreePicker(id=NULL, appearance=NULL, block=NULL, cascade=NULL,
cleanable=NULL, countable=NULL, custom_locale=NULL,
data=NULL, defaultExpandAll=NULL, placement=NULL,
selected=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{appearance}{Character. The appearance of the component. Can be `default` or `subtle`.}

\item{block}{Logical. Whether to take up the full width of the parent container.}

\item{cascade}{Logical. Whether to enable cascade selection.}

\item{cleanable}{Logical. Whether to display the clear button}

\item{countable}{Logical. Whether to display the count of selected items}

\item{custom_locale}{Named list. Custom locale for component}

\item{data}{Unnamed list. The Tree Data}

\item{defaultExpandAll}{Logical. Whether to expand all nodes by default}

\item{placement}{Character. The placement of the popup. Can be 'bottomStart', 'bottomEnd', 'topStart', 'topEnd'.}

\item{selected}{List of characters. Selected value(s)}

\item{style}{Named list. Custom styling of the component.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

